Read search optimization: request only the fields we use
This omits, for example, the unused alignedBases and tags field of reads which
can be quite large.  In one example of navigating a 1000 genomes readset at
the 'reads' scale, this reduced data transfer volume for a typical navigation
from 400kB to 150kB, and client load time from about 2s to 1.5s.