Use fixed-height tracks stacked vertically
Rather than space the tracks out into the available space, stack them up along
the bottom with a fixed height each.  This is the first tiny step in a larger UI
overhaul to enable fluid navigation.  Once a given read is in view it should
remain in the same relative position when panning left and right.