Initial cut at fullscreen mobile-optimized UI
This overhauls the UI according to our mobile-first mocks.  This is still very
rough and supports only the graph view (must use a URL hash to test as there's
no UI for selecting backends / reads).

One key change is that the UI is now full screen and esponsive to size changes,
adapting on window resize (including device rotation).